requirements.txt
# HeadElf Python Dependencies
# Core execution dependencies for Claude Code integration

# Core Python libraries
asyncio
pathlib
json
typing
logging
subprocess

# YAML processing for metadata
PyYAML>=6.0

# JSON schema validation
jsonschema>=4.17.0

# HTTP requests for API integrations
requests>=2.28.0

# Date/time handling
python-dateutil>=2.8.2

# Configuration management
configparser>=5.3.0

# Testing framework
pytest>=7.2.0
pytest-asyncio>=0.21.0

# Development dependencies
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Code linting
mypy>=1.0.0    # Type checking

# Optional dependencies for enhanced features
# Uncomment if advanced features are needed

# Database support (optional)
# sqlalchemy>=2.0.0
# sqlite3  # Built into Python

# Data analysis (optional)
# pandas>=2.0.0
# numpy>=1.24.0

# Visualization (optional)
# matplotlib>=3.7.0
# plotly>=5.13.0

# Enterprise integrations (optional)
# ldap3>=2.9.1        # LDAP authentication
# cryptography>=40.0.0 # Encryption support
# jwt>=1.3.1          # JWT token handling

# Cloud integrations (optional)
# boto3>=1.26.0       # AWS integration
# azure-identity>=1.12.0 # Azure integration
# google-cloud>=0.34.0   # Google Cloud integration

# Monitoring and observability (optional)
# prometheus-client>=0.16.0
# opentelemetry-api>=1.17.0